+++ .env
@@ -1,0 +1,2 @@
+NODE_PATH=src/
+
