+++ src/Pages/Landing/Landing.js
@@ -5,7 +5,7 @@ import "antd/dist/antd.css";
 import HomePage from "./HomePage/HomePage.js";
 import SponsorPage from "./SponsorPage/SponsorPage.js";
 import AboutPage from "./AboutPage/AboutPage.js";
-import Footer from "./Footer/Footer.js";
+import Footer from "CommonComponents/Footer/Footer.js";
 
 const mobileThreshold = 900;
