+++ src/Pages/Landing/SponsorPage/SponsorPage.js
@@ -28,12 +28,12 @@ class SponsorPage extends React.Component {
           </Flex>
           <SponsorContainer>
             <SponsorRow {...{ isWebView }} >
-              <SponsorLogoBig {...{ isWebView }} src={require("../../assets/SponsorLogos/AA.png")} />
-              <SponsorLogoBig {...{ isWebView }} src={require("../../assets/SponsorLogos/charles.png")} />
+              <SponsorLogoBig {...{ isWebView }} src={require("assets/SponsorLogos/AA.png")} />
+              <SponsorLogoBig {...{ isWebView }} src={require("assets/SponsorLogos/charles.png")} />
             </SponsorRow>
             <SponsorRow {...{ isWebView }}>
-              <SponsorLogo {...{ isWebView }} src={require("../../assets/SponsorLogos/exxonmobile.png")} />
-              <SponsorLogo {...{ isWebView }} src={require("../../assets/SponsorLogos/google.png")} />
+              <SponsorLogo {...{ isWebView }} src={require("assets/SponsorLogos/exxonmobile.png")} />
+              <SponsorLogo {...{ isWebView }} src={require("assets/SponsorLogos/google.png")} />
             </SponsorRow>
           </SponsorContainer>
           <Flex style={{ justifyContent: "center", fontSize: "14px", flexDirection: "column" }}>
@@ -46,27 +46,27 @@ class SponsorPage extends React.Component {
           </Flex>
           <SponsorContainer>
             <SponsorRow {...{ isWebView }}>
-              <SponsorLogoBig {...{ isWebView }} src={require("../../assets/SponsorLogos/AA.svg")} />
+              <SponsorLogoBig {...{ isWebView }} src={require("assets/SponsorLogos/AA.svg")} />
             </SponsorRow>
             <SponsorRow {...{ isWebView }}>
-              <SponsorLogo {...{ isWebView }} src={require("../../assets/SponsorLogos/chase.png")} />
-              <SponsorLogo {...{ isWebView }} src={require("../../assets/SponsorLogos/google.png")} />
-              <SponsorLogo {...{ isWebView }} src={require("../../assets/SponsorLogos/microsoft.png")} />
+              <SponsorLogo {...{ isWebView }} src={require("assets/SponsorLogos/chase.png")} />
+              <SponsorLogo {...{ isWebView }} src={require("assets/SponsorLogos/google.png")} />
+              <SponsorLogo {...{ isWebView }} src={require("assets/SponsorLogos/microsoft.png")} />
             </SponsorRow>
             <SponsorRow {...{ isWebView }}>
-              <SponsorLogo {...{ isWebView }} src={require("../../assets/SponsorLogos/southwest.png")} />
-              <SponsorLogo {...{ isWebView }} src={require("../../assets/SponsorLogos/facebook.png")} />
-              <SponsorLogo {...{ isWebView }} src={require("../../assets/SponsorLogos/capital_one.png")} />
+              <SponsorLogo {...{ isWebView }} src={require("assets/SponsorLogos/southwest.png")} />
+              <SponsorLogo {...{ isWebView }} src={require("assets/SponsorLogos/facebook.png")} />
+              <SponsorLogo {...{ isWebView }} src={require("assets/SponsorLogos/capital_one.png")} />
             </SponsorRow>
             <SponsorRow {...{ isWebView }}>
-              <SponsorLogo {...{ isWebView }} src={require("../../assets/SponsorLogos/accenture.png")} />
-              <SponsorLogo {...{ isWebView }} src={require("../../assets/SponsorLogos/goldman.png")} />
-              <SponsorLogo {...{ isWebView }} src={require("../../assets/SponsorLogos/esri.png")} />
+              <SponsorLogo {...{ isWebView }} src={require("assets/SponsorLogos/accenture.png")} />
+              <SponsorLogo {...{ isWebView }} src={require("assets/SponsorLogos/goldman.png")} />
+              <SponsorLogo {...{ isWebView }} src={require("assets/SponsorLogos/esri.png")} />
             </SponsorRow>
             <SponsorRow {...{ isWebView }}>
-              <SponsorLogo {...{ isWebView }} src={require("../../assets/SponsorLogos/hbk.png")} />
-              <SponsorLogo {...{ isWebView }} src={require("../../assets/SponsorLogos/pariveda.png")} />
-              <SponsorLogo {...{ isWebView }} src={require("../../assets/SponsorLogos/gmnew.png")} />
+              <SponsorLogo {...{ isWebView }} src={require("assets/SponsorLogos/hbk.png")} />
+              <SponsorLogo {...{ isWebView }} src={require("assets/SponsorLogos/pariveda.png")} />
+              <SponsorLogo {...{ isWebView }} src={require("assets/SponsorLogos/gmnew.png")} />
             </SponsorRow>
           </SponsorContainer>
         </Flex>
