+++ src/index.js
@@ -1,9 +1,9 @@
 /* eslint-disable */
 import React from "react";
 import ReactDOM from "react-dom";
-import AppRouter from "./AppRouter";
+import PageRouter from "./Pages/PageRouter";
 import registerServiceWorker from "./registerServiceWorker";
 import "./index.css";
 
-ReactDOM.render(<AppRouter />, document.getElementById("root"));
+ReactDOM.render(<PageRouter />, document.getElementById("root"));
 registerServiceWorker();
